Totally reorganized code
