+++ utils/llm/past_conversations.py
@@ -1,0 +1,1 @@
+history = []
